\name{bergerPeriods}
\alias{bergerPeriods}
\title{Obliquity and precession periods of Berger et al. (1992)}
\description{
Determine the predicted precession and obliquity periods based on Berger et al. (1992). Values are determined by piecewise linear interpolation.
}
\usage{
bergerPeriods(age,genplot=T)
}
\arguments{
  \item{age}{Age (millions of years before present)}
  \item{genplot}{Generate summary plots? (T or F)}
}

\references{
A. Berger, M.F. Loutre, and J. Laskar, 1992,
\emph{Stability of the Astronomical Frequencies Over the Earth's History for Paleoclimate Studies}: Science, v. 255, p. 560-566.
}