\name{laskar} 
\docType{data} 
\alias{laskar} 
\title{Eccentricity, Tilt and Precession Solutions for the Past 250 Million Years} 
\description{ 
   Astronomical solutions from Laskar: Precession, obliquity, and eccentricity solutions from Laskar et al. (2004), and the four eccentricity solutions from Laskar et al. (2011). 
} 
\usage{laskar} 
\format{Time (ka before present, from J2000), precession angle (longitude of perihelion from moving equinox, in radians), obliquity (radians), eccentricity (LA10a), eccentricity (LA10b), eccentricity (LA10c), eccentricity (LA10d), eccentricity (LA04)} 
\source{http://www.imcce.fr/Equipes/ASD/insola/earth/earth.html} 

\references{ 
Laskar, J., Robutel, P., Joutel, F., Gastineau, M., Correia, A.C.M., Levrard, B., 2004, 
\emph{A long term numerical solution for the insolation quantities of the Earth}: Astron. Astrophys., Volume 428, 261-285.

Laskar, J., Fienga, A., Gastineau, M., Manche, H., 2011, 
\emph{La2010: A new orbital solution for the long-term motion of the Earth}: Astron. Astrophys., Volume 532, A89.
}
