\name{plS}
\alias{plS}
\title{Set default plotting parameters for vertical stratigraphic plots}
\description{
Set default plotting parameters for vertical stratigraphic plots. This is ususally invoked after function pl.
}
\usage{
plS(f=T,s=1)
}
\arguments{
  \item{f}{Are you plotting the first (leftmost) stratigraphic plot? (T or F)}
  \item{s}{Size of the symbols and text on plot. Default = 1}
}
