\name{rmNA}
\alias{rmNA}
\title{Remove stratigraphic levels that contain one or more NAs}
\description{
Remove stratigraphic levels that contain one or more NAs.
 
}
\usage{
rmNA(dat,genplot=T,verbose=T)
}
\arguments{
  \item{dat}{Data series to process. If location is included (e.g., depth), it should be in the first column.}
  \item{genplot}{Generate summary plots? (T or F)}
  \item{verbose}{Verbose output? (T or F)}
}