\name{strats}
\alias{strats}
\title{Summary Statistics for Stratigraphic Series}
\description{
Summary statistics for stratigraphic series: sampling interval and proxy values.
}
\usage{
strats(dat)
}
\arguments{
  \item{dat}{Stratigraphic series to evaluate. First column should be location (e.g., depth), second column should be data value.}
}