\name{xplot}
\alias{xplot}
\title{Generate Cross-plot with Density Estimates on Axes}
\description{
Generate a Cross-plot with Density Estimates on Axes. Custom axes titles optional.
}
\usage{
xplot(x,y,xlab=NULL,ylab=NULL,main=NULL)
}
\arguments{
  \item{x}{Variable 1}
  \item{y}{Variable 2}
  \item{xlab}{Label for the x-axis, in quotes}
  \item{ylab}{Label for the y-axis, in quotes}
  \item{main}{Label for the plot, in quotes}
}