\name{getData}
\alias{getData}
\title{Download file from astrochron data server}
\description{
Download data file from astrochron server.
}
\usage{
getData(dat="1262-a*")
}

\arguments{
  \item{dat}{A character string that specifies the data file to download. At present there are two options: "1262-a*" or "926B-18O"}
}